work and guest creation flow

+++ db/migrate/20130513062936_create_works.rb
@@ -3,7 +3,7 @@ class CreateWorks < ActiveRecord::Migration
     create_table :works do |t|
       t.string :name
       t.string :client
-      t.decimal :year
+      t.integer :year
       t.string :role
       t.text :description
       t.boolean :private_work
